backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.1
python-dotenv==1.0.0
